+++ packages/react/src/experimental/Forms/Form/index.stories.tsx
@@ -1,8 +1,8 @@
 import type { Meta, StoryObj } from "@storybook/react-vite"
 
+import { Button } from "@/components/Actions/Button"
 import { UseFormReturn } from "react-hook-form"
 import { number, z } from "zod"
-import { Button } from "../../../factorial-one"
 import { Textarea } from "../../../ui/textarea"
 import { AutoGrid } from "../../Utilities/Layout/AutoGrid"
 import { Checkbox } from "../Fields/Checkbox"