+++ packages/react/src/experimental/OneDataCollection/visualizations/collection/collectionViewRegistry.tsx
@@ -1,13 +1,14 @@
+import { FiltersDefinition } from "@/components/OneFilterPicker"
 import { IconType } from "@/components/Utilities/Icon"
+import { ItemActionsDefinition } from "@/experimental/OneDataCollection/item-actions"
+import { NavigationFiltersDefinition } from "@/experimental/OneDataCollection/navigationFilters/types"
 import { Kanban, List, Table } from "@/icons/app"
 import {
-  FiltersDefinition,
-  ItemActionsDefinition,
-  NavigationFiltersDefinition,
+  GroupingDefinition,
+  RecordType,
   SortingsDefinition,
   SummariesDefinition,
-} from "../../exports"
-import { GroupingDefinition, RecordType } from "../../types"
+} from "../../types"
 import { CardCollection, CardCollectionProps } from "./Card"
 import { ListCollection, ListCollectionProps } from "./List"
 import { TableCollection, TableCollectionProps } from "./Table"