+++ packages/react/src/experimental/exports.ts
@@ -1,6 +1,7 @@
 import { Component } from "../lib/component"
 import { ScrollArea as ScrollAreaComponent } from "./Utilities/ScrollArea"
 
+export * from "./AiChat"
 export * from "./Banners/exports"
 export * from "./Charts/exports"
 export * from "./Forms/exports"